tasks-api/app/requirements.txt
# Web framework
fastapi==0.68.0
uvicorn==0.15.0

# Database
psycopg2-binary==2.9.1

# AWS
boto3==1.18.65

# File handling
python-multipart==0.0.5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==1.8.2

# HTTP client (for testing)
httpx==0.19.0

# Environment variables
python-dotenv==0.19.0

# Logging
logging==0.4.9.6

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1

# CORS
starlette==0.14.2

# Mangum for AWS Lambda integration
mangum==0.12.0